extracted CustomHeader save functions
